fix : jpa를 이용하여 상품 페이징하여 보여주도록 변경

+++ week5/kakao-5th-exhandler-start/src/main/java/com/example/kakao/product/domain/service/ProductRepository.java
@@ -9,6 +9,5 @@ import java.util.Optional;
 public interface ProductRepository {
     Optional<ProductEntity> findById(Long id);
 
-    List<ProductEntity> findAll(Long page);
-
+    List<ProductEntity> findAll(int page);
 }
